#22 메인슬라이드배너 관리 추가중,이벤트컨트롤러 분리-dh-
